:truck: Renamed _entity.py to _bionty.py
