+++ bionty/__init__.py
@@ -82,7 +82,7 @@ from .entities._readout import Readout
 
 # tools
 from ._normalize import NormalizeColumns
-from ._entity import Bionty
+from ._bionty import Bionty
 from ._ontology import Ontology
 from ._display_sources import display_currently_used_sources, display_available_sources
