+++ bionty/entities/_cellline.py
@@ -1,6 +1,6 @@
 from typing import Literal, Optional
 
-from .._entity import Bionty
+from .._bionty import Bionty
 from ._shared_docstrings import _doc_params, doc_entites
