+++ bionty/entities/_protein.py
@@ -2,7 +2,7 @@ from typing import Literal, Optional
 
 import pandas as pd
 
-from .._entity import Bionty
+from .._bionty import Bionty
 from .._normalize import NormalizeColumns
 from ..dev._io import s3_bionty_assets
 from ._shared_docstrings import _doc_params, doc_entites
